pose estimation

+++ controller.py
@@ -426,7 +426,7 @@ class Controller:
 
         for j in range(1):
             for point in points:
-                for i in range(self.flight_duration * 10):
+                for i in range(int(self.flight_duration * 10)):
                     if self.battery_low:
                         return
                     self.send_position_target(point[0], point[1], 0)
